moving some of the commands to an external file to improve searching for robustness
